refactor: 회원가입 로직을 user 도메인으로 변경
