+++ src/main/java/com/cliptripbe/feature/user/application/UserService.java
@@ -1,10 +1,15 @@
 package com.cliptripbe.feature.user.application;
 
+import com.cliptripbe.feature.user.domain.entity.User;
+import com.cliptripbe.feature.user.domain.service.UserChecker;
 import com.cliptripbe.feature.user.domain.service.UserLoader;
+import com.cliptripbe.feature.user.dto.request.UserSignUpRequest;
 import com.cliptripbe.feature.user.dto.response.UserInfoResponse;
+import com.cliptripbe.feature.user.infrastructure.UserRepository;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -14,6 +19,18 @@ import org.springframework.transaction.annotation.Transactional;
 public class UserService {
 
     private final UserLoader userLoader;
+    private final PasswordEncoder passwordEncoder;
+    private final UserChecker userChecker;
+    private final UserRepository userRepository;
+
+    @Transactional
+    public UserInfoResponse signUp(UserSignUpRequest signUpDto) {
+        userChecker.checkExistUserEmail(signUpDto.email());
+        String encodePassword = passwordEncoder.encode(signUpDto.password());
+        User user = signUpDto.toEntity(encodePassword);
+        userRepository.save(user);
+        return UserInfoResponse.from(user);
+    }
 
     public List<UserInfoResponse> getAllUserInfo() {
         return userLoader.getAllUser()
